% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cjointPlot.R
\docType{package}
\name{cjointPlot}
\alias{cjointPlot}
\alias{cjointPlot-package}
\title{amceオブジェクトのプロッティング関数}
\arguments{
\item{cjoint.obj}{: 'amce' class object}

\item{show.est}{: Show a estimated coefficients. Default is TRUE}

\item{show.guide}{: Show a guidelines. Default is TRUE}

\item{xlab}{: Customize x-axis label. Default is NULL}

\item{title}{: Customise a plot title. Default is NULL}
}
\value{
A class of a return value is "NULL"
}
\description{
\code{cjointPlot}は\code{cjoint}パッケージから生成されたamceオブジェクトを引数とし、推定値のプロットを表示します。
\code{cjoint}内蔵の\code{plot}メソッドとは違い、多様な言語に対応し、より軽量化を図りました。
}
\examples{
cjoint.model <- amce(selected ~ X1 + X2 + X3, data = choice4to2.df, respondent.id = "respondentIndex", cluster = TRUE, design = cjoint.design)
cjointPlot(cjoint.model)
cjointPlot(cjoint.model, show.est = FALSE, title = "Test Plot")
}
\author{
SONG Jaehyun
(jaehyun.song [at] stu.kobe-u.ac.jp ||
http://www.JaySong.net)
}

